Created seperate component for Room types + Datepicker error changes
