+++ src/store/DatePickerStore.ts
@@ -1,4 +1,4 @@
-import { DateError, DateValue } from "@/components/datePicker";
+import { DateValue } from "@/components/datePicker";
 import { create } from "zustand";
 
 interface DatePicker {
@@ -6,18 +6,11 @@ interface DatePicker {
     setCheckIn: (checkInDate: DateValue) => void;
     checkOut: DateValue;
     setCheckOut: (checkOutDate: DateValue) => void;
-    dateError: DateError;
-    setDateError: (dateError: DateError) => void
 }
 
 export const useDatePickerStore = create<DatePicker>()((set) => ({
     checkIn: null,
     setCheckIn: (checkInDate) => set({ checkIn: checkInDate }),
     checkOut: null,
-    setCheckOut: (checkOutDate) => set({ checkOut: checkOutDate }),
-    dateError: {
-        checkInError: false,
-        checkOutError: false
-    },
-    setDateError: (dateError) => set({ dateError: dateError })
+    setCheckOut: (checkOutDate) => set({ checkOut: checkOutDate })
 }));
